add a flag to control whether delete data of dropped space when restart (#2375)
* add a flag to control whether delete data of dropped space when restart

* address @darionyaphet's comments

+++ src/kvstore/NebulaStore.cpp
@@ -21,6 +21,7 @@ DEFINE_int32(custom_filter_interval_secs, 24 * 3600,
 DEFINE_int32(num_workers, 4, "Number of worker threads");
 DEFINE_bool(check_leader, true, "Check leader or not");
 DEFINE_int32(clean_wal_interval_secs, 600, "inerval to trigger clean expired wal");
+DEFINE_bool(auto_remove_invalid_space, false, "whether remove data of invalid space when restart");
 
 DECLARE_bool(rocksdb_disable_wal);
 
@@ -72,14 +73,16 @@ bool NebulaStore::init() {
                     }
 
                     if (!options_.partMan_->spaceExist(storeSvcAddr_, spaceId).ok()) {
-                        // TODO We might want to have a second thought here.
-                        // Removing the data directly feels a little strong
-                        LOG(INFO) << "Space " << spaceId
-                                  << " does not exist any more, remove the data!";
-                        auto dataPath = folly::stringPrintf("%s/%s",
-                                                            rootPath.c_str(),
-                                                            dir.c_str());
-                        CHECK(fs::FileUtils::remove(dataPath.c_str(), true));
+                        if (FLAGS_auto_remove_invalid_space) {
+                            // TODO We might want to have a second thought here.
+                            // Removing the data directly feels a little strong
+                            LOG(INFO) << "Space " << spaceId
+                                      << " does not exist any more, remove the data!";
+                            auto dataPath = folly::stringPrintf("%s/%s",
+                                                                rootPath.c_str(),
+                                                                dir.c_str());
+                            CHECK(fs::FileUtils::remove(dataPath.c_str(), true));
+                        }
                         continue;
                     }
